+++ CHANGELOG.md
@@ -3,11 +3,14 @@
 ## Unreleased (3.6)
 
 ### Added
+- Entries now begin life as “unpublished drafts” rather than “unsaved drafts”. They are no longer ephemeral; they will continue to exist until explicitly published or deleted. ([#5661](https://github.com/craftcms/cms/issues/5661), [#7216](https://github.com/craftcms/cms/issues/7216))
 - User indexes can now include a “Groups” column. ([#7211](https://github.com/craftcms/cms/issues/7211))
 - It’s now possible to set sites’ Name settings to environment variables.
 - Added the `{% tag %}` Twig tag.
 - Added the `withGroups` user query param.
 - Added the `relatedToAssets`, `relatedToCategories`, `relatedToEntries`, `relatedToTags`, and `relatedToUsers` arguments to GraphQL queries. ([#7110](https://github.com/craftcms/cms/issues/7110))
+- Added the `isUnpublishedDraft` GraphQL field.
+- Added `craft\base\ElementInterface::getIsUnpublishedDraft()`.
 - Added `craft\events\RegisterGqlArgumentHandlersEvent`.
 - Added `craft\gql\ArgumentManager`.
 - Added `craft\gql\base\ArgumentHandler`.
@@ -27,6 +30,9 @@
 - Added `craft\models\Site::getName()`.
 - Added `craft\models\Site::setBaseUrl()`.
 - Added `craft\models\Site::setName()`.
+- Added `craft\services\Drafts::EVENT_AFTER_APPLY_DRAFT`.
+- Added `craft\services\Drafts::EVENT_BEFORE_APPLY_DRAFT`.
+- Added `craft\services\Drafts::publishDraft()`.
 
 ### Changed
 - `craft\behaviors\EnvAttributeParserBehavior::$attributes` can now be set to an array with key/value pairs, where the key is the attribute name, and the value is the raw (unparsed) value, or a callable that returns the raw value.
@@ -35,10 +41,17 @@
 - `craft\models\Site::getBaseUrl()` now has a `$parse` argument, which can be set to `false` to return the raw (unparsed) base URL.
 
 ### Deprecated
+- Deprecated the `purgeUnsavedDraftsDuration` config setting.
 - Deprecated the `siteName` config setting. Sites’ Name settings should be set to environment variables instead.
 - Deprecated the `siteUrl` config setting. Sites’ Base URL settings should be set to aliases or environment variables instead. ([#3205](https://github.com/craftcms/cms/issues/3205))
+- Deprecated the `isUnsavedDraft` GraphQL field.
+- Deprecated `craft\base\Element::getIsUnsavedDraft()`. `getIsUnpublishedDraft()` should be used instead.
 - Deprecated `craft\gql\base\Resolver::getArrayableArguments()`.
 - Deprecated `craft\gql\base\Resolver::prepareArguments()`. 
+- Deprecated `craft\services\Drafts::applyDraft()`. `publishDraft()` should be used instead.
+- Deprecated `craft\services\Drafts::EVENT_AFTER_APPLY_DRAFT`. `EVENT_AFTER_PUBLISH_DRAFT` should be used instead.
+- Deprecated `craft\services\Drafts::EVENT_BEFORE_APPLY_DRAFT`. `EVENT_BEFORE_PUBLISH_DRAFT` should be used instead.
+- Deprecated `craft\services\Drafts::purgeUnsavedDrafts()`.
 
 ### Fixed
 - Fixed a JavaScript error that occurred when opening an element selector modal. ([#7186](https://github.com/craftcms/cms/issues/7186))